services/ocr-service/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.9
pytesseract>=0.3.10
pdf2image>=1.17.0
Pillow>=10.2.0
opencv-python-headless>=4.9.0
structlog>=24.1.0
pydantic-settings>=2.1.0
pydantic>=2.6.0
python-dotenv>=1.0.1
# Test Dependencies
pytest>=8.0.0
httpx>=0.26.0
requests>=2.31.0
